11_application_code/requirements-api.txt
# TFOS API runtime dependencies — production image
# Strictly what the running app/ code imports. Anything else belongs in
# requirements-dev.txt or in a service-specific requirements file.

# Web
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17

# Database
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
pgvector==0.3.6

# Cache + Queue
redis==5.2.1
celery[redis]==5.4.0
kombu==5.4.2

# AI
anthropic==0.40.0
openai==1.57.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# HTTP (also used for Resend API calls — no SDK needed)
httpx==0.27.2

# Config
python-dotenv==1.0.1

# Validation
pydantic[email]==2.10.3
pydantic-settings==2.6.1

# Payments
stripe==11.3.0

# Monitoring
sentry-sdk[fastapi]==2.19.2
structlog==24.4.0

# Utils
python-dateutil==2.9.0
pytz==2024.2
aiofiles==24.1.0
tenacity==9.0.0
orjson==3.10.12
